Move functions used by script into submodule
